# docs/api/openapi-gamification.yaml
openapi: 3.0.3
info:
  title: DevSocial Gamification API
  version: 1.0.0
  description: Gamification service for XP, badges, challenges, and leaderboards

servers:
  - url: http://localhost:3000/api
    description: Development
  - url: https://devsocial.app/api
    description: Production

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
        username:
          type: string
        displayName:
          type: string
        avatar:
          type: string
        level:
          type: integer

    LeaderboardEntry:
      type: object
      properties:
        _id:
          type: string
        user:
          $ref: '#/components/schemas/User'
        totalXP:
          type: integer
        rank:
          type: integer
        level:
          type: integer
        referralCount:
          type: integer
          nullable: true
        challengesCompleted:
          type: integer
          nullable: true

    Challenge:
      type: object
      properties:
        _id:
          type: string
        title:
          type: string
        description:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        xpReward:
          type: integer
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        status:
          type: string
          enum: [active, completed, expired]

paths:
  /gamification/award-xp:
    post:
      summary: Award XP to user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - action
              properties:
                action:
                  type: string
                  enum: [post_created, comment_created, post_liked, daily_login, challenge_completed]
                content:
                  type: string
                refId:
                  type: string
      responses:
        '200':
          description: XP awarded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      xpAwarded:
                        type: integer
                      newTotal:
                        type: integer
                      levelUp:
                        type: boolean
        '401':
          description: Unauthorized
        '400':
          description: Bad request

  /leaderboard:
    get:
      summary: Get leaderboard
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [all-time, weekly, monthly, referrals, challenges]
            default: all-time
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Leaderboard data
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      leaderboard:
                        type: array
                        items:
                          $ref: '#/components/schemas/LeaderboardEntry'
                      type:
                        type: string
                      totalUsers:
                        type: integer

  /challenges:
    get:
      summary: Get active challenges
      responses:
        '200':
          description: List of challenges
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      challenges:
                        type: array
                        items:
                          $ref: '#/components/schemas/Challenge'
    post:
      summary: Create challenge (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - description
                - difficulty
                - xpReward
              properties:
                title:
                  type: string
                description:
                  type: string
                difficulty:
                  type: string
                  enum: [easy, medium, hard]
                xpReward:
                  type: integer
                startDate:
                  type: string
                  format: date-time
                endDate:
                  type: string
                  format: date-time
      responses:
        '201':
          description: Challenge created
        '403':
          description: Forbidden - Admin only

  /challenges/{challengeId}/join:
    post:
      summary: Join a challenge
      security:
        - bearerAuth: []
      parameters:
        - name: challengeId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Joined successfully
        '401':
          description: Unauthorized

  /challenges/{challengeId}/submit:
    post:
      summary: Submit challenge solution
      security:
        - bearerAuth: []
      parameters:
        - name: challengeId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                solution:
                  type: string
                code:
                  type: string
      responses:
        '200':
          description: Submission received
        '401':
          description: Unauthorized